Extract `:logging-dynamic-backend` module
